fix: Convert XLIFF to JSON files in proper directory and with optional prolog
* Place the translated i18n.json files to the directory named according to ISO 639-3. The run-time package (vscode-nls) expects them there.
* Do not prepend Microsoft copyright text as a prolog for the i18n.json files by default. Allow the prolog to be prepended by an optional parameter.
* Allow converting XLIFF files with no translations to templates for i18n.json files.

BREAKING CHANGE: The Gulp task `prepareJsonFiles` requires the `languages` parameter to process XLIFF files with translations and the `prolog` parameter with the Microsoft copyright text if you want it to behave like it did before. But the translated i18n.json files will be always stored in a directory named according to ISO 639-3 instead of directly in the output directory.

When storing the translated i18n.json files, the directory name used for the target language has to be known. When storing the translated i18n.json files, the parameter `languages` for `prepareJsonFiles` is required. It does not make sense supporting the previous functionality by optionally omitting the language-specific sub-directory, because the run-time package would not work with it.

Not all projects are governed by Microsoft. The prolog for the i18n.json files should be optional.